LayoutTests/fast/dom/DeviceMotion/optional-event-properties-expected.txt

Tests the optional properties of DeviceMotionEvent. Each property should be null if not set, or set to null or undefined.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


event = document.createEvent('DeviceMotionEvent')
PASS event.xAcceleration == null is true
PASS event.yAcceleration == null is true
PASS event.zAcceleration == null is true
PASS event.xRotationRate == null is true
PASS event.yRotationRate == null is true
PASS event.zRotationRate == null is true
PASS event.interval == null is true
event.initDeviceMotionEvent('', false, false, 0, 1, 2, 3, 4, 5, 6)
PASS event.xAcceleration == 0 is true
PASS event.yAcceleration == 1 is true
PASS event.zAcceleration == 2 is true
PASS event.xRotationRate == 3 is true
PASS event.yRotationRate == 4 is true
PASS event.zRotationRate == 5 is true
PASS event.interval == 6 is true
event.initDeviceMotionEvent()
PASS event.xAcceleration == null is true
PASS event.yAcceleration == null is true
PASS event.zAcceleration == null is true
PASS event.xRotationRate == null is true
PASS event.yRotationRate == null is true
PASS event.zRotationRate == null is true
PASS event.interval == null is true
event.initDeviceMotionEvent('', false, false, [], [], [], [], [], [], [])
PASS event.xAcceleration == 0 is true
PASS event.yAcceleration == 0 is true
PASS event.zAcceleration == 0 is true
PASS event.xRotationRate == 0 is true
PASS event.yRotationRate == 0 is true
PASS event.zRotationRate == 0 is true
PASS event.interval == 0 is true
event.initDeviceMotionEvent('', false, false, undefined, undefined, undefined, undefined, undefined, undefined, undefined)
PASS event.xAcceleration == null is true
PASS event.yAcceleration == null is true
PASS event.zAcceleration == null is true
PASS event.xRotationRate == null is true
PASS event.yRotationRate == null is true
PASS event.zRotationRate == null is true
PASS event.interval == null is true
event.initDeviceMotionEvent('', false, false, '', '', '', '', '', '', '')
PASS event.xAcceleration == 0 is true
PASS event.yAcceleration == 0 is true
PASS event.zAcceleration == 0 is true
PASS event.xRotationRate == 0 is true
PASS event.yRotationRate == 0 is true
PASS event.zRotationRate == 0 is true
PASS event.interval == 0 is true
event.initDeviceMotionEvent('', false, false, null, null, null, null, null, null, null)
PASS event.xAcceleration == null is true
PASS event.yAcceleration == null is true
PASS event.zAcceleration == null is true
PASS event.xRotationRate == null is true
PASS event.yRotationRate == null is true
PASS event.zRotationRate == null is true
PASS event.interval == null is true
PASS successfullyParsed is true

TEST COMPLETE
